Programming Personal Project Commit #1
In this commit I’m uploading code that I did not initially write. This
code generates a waveform and plays it back at runtime. The only
modifications I’ve made are in OscillatorPlayer.java, I took out
everything that referenced gnu.getopt to see if it would affect the
final result of the code. No noticeable changes ocurred.